+++ FitFinder/FitFinder.xcodeproj/xcuserdata/noahfrew.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "4C1756DB-97D3-4DBE-9F1B-5CA61D9EFDB3"
+   type = "1"
+   version = "2.0">
+</Bucket>
